+++ casper/src/test/scala/coop/rchain/casper/batch1/MultiParentCasperRholangSpec.scala
@@ -1,10 +1,9 @@
 package coop.rchain.casper.batch1
 
-import coop.rchain.casper.BlockRandomSeed
 import coop.rchain.casper.helper.TestNode
 import coop.rchain.casper.helper.TestNode.Effect
 import coop.rchain.casper.protocol.BlockMessage
-import coop.rchain.casper.rholang.{RuntimeManager, Tools}
+import coop.rchain.casper.rholang.{BlockRandomSeed, RuntimeManager, Tools}
 import coop.rchain.casper.util.{ConstructDeploy, ProtoUtil, RSpaceUtil}
 import coop.rchain.crypto.signatures.Secp256k1
 import coop.rchain.p2p.EffectsTestInstances.LogicalTime
