+++ casper/src/test/scala/coop/rchain/casper/helper/RhoSpec.scala
@@ -2,13 +2,12 @@ package coop.rchain.casper.helper
 
 import cats.effect.{Concurrent, Sync}
 import cats.syntax.all._
-import coop.rchain.casper.BlockRandomSeed
 import coop.rchain.casper.genesis.Genesis
 import coop.rchain.casper.genesis.contracts.TestUtil
 import coop.rchain.casper.genesis.contracts.TestUtil.eval
 import coop.rchain.casper.protocol.DeployData
 import coop.rchain.casper.rholang.Resources.{copyStorage, mkTestRNodeStoreManager}
-import coop.rchain.casper.rholang.{Resources, Tools}
+import coop.rchain.casper.rholang.{BlockRandomSeed, Resources, Tools}
 import coop.rchain.casper.util.GenesisBuilder
 import coop.rchain.casper.util.GenesisBuilder.GenesisParameters
 import coop.rchain.crypto.PrivateKey
