+++ node/src/main/scala/coop/rchain/node/revvaultexport/StateBalances.scala
@@ -5,8 +5,8 @@ import cats.effect.{Concurrent, ContextShift, Sync}
 import cats.syntax.all._
 import com.google.protobuf.ByteString
 import coop.rchain.blockstorage.BlockStore
-import coop.rchain.casper.BlockRandomSeed
 import coop.rchain.casper.genesis.Genesis
+import coop.rchain.casper.rholang.BlockRandomSeed
 import coop.rchain.casper.rholang.RuntimeManager.emptyStateHashFixed
 import coop.rchain.casper.storage.RNodeKeyValueStoreManager
 import coop.rchain.metrics.{Metrics, NoopSpan}
