+++ node/src/main/scala/coop/rchain/node/runtime/Setup.scala
@@ -12,10 +12,9 @@ import coop.rchain.casper.blocks.proposer.{Proposer, ProposerResult}
 import coop.rchain.casper.blocks.{BlockProcessor, BlockReceiver, BlockReceiverState, BlockRetriever}
 import coop.rchain.casper.dag.BlockDagKeyValueStorage
 import coop.rchain.casper.engine.{NodeLaunch, PeerMessage}
-import coop.rchain.casper.genesis.Genesis
 import coop.rchain.casper.protocol.{toCasperMessageProto, BlockMessage, CasperMessage, CommUtil}
 import coop.rchain.casper.reporting.{ReportStore, ReportingCasper}
-import coop.rchain.casper.rholang.RuntimeManager
+import coop.rchain.casper.rholang.{BlockRandomSeed, RuntimeManager}
 import coop.rchain.casper.state.instances.{BlockStateManagerImpl, ProposerState}
 import coop.rchain.casper.syntax._
 import coop.rchain.comm.RoutingMessage
